+++ __tests__/test-mock-data/capabilities.ts
@@ -1,0 +1,7 @@
+import objectBuilder from './object-builder';
+
+export const mockCapability = objectBuilder({
+  name: 'test_capability',
+  status: 'ACTIVE',
+  description: 'Capability description',
+});